fix(android): resolve build errors; keep OrbitX red branding and Infrawire sponsor intact
- Fix the module rename references (:orbitx_launcher) across all Gradle scripts.
- Repair the class-name collision introduced by the branding sweep in
  sponsor/InfrawirePartner* and the launcher's sponsor-dialog call sites.
- FirebaseSyncManager: drop the Markdown.toPlain() dependency.
- ScreenRecorder: explicit int casts for the capture scale maths.
- Add the missing default strings (login_microsoft, login_error_invalid_username).

+++ android/orbitx_launcher/src/main/java/net/kdt/pojavlaunch/recorder/ScreenRecorder.java
@@ -106,8 +106,8 @@ public class ScreenRecorder {
             int height = size.y;
             int longEdge = Math.max(width, height);
             float scale = longEdge > 1920 ? 1920f / longEdge : 1f;
-            int encWidth = even(width * scale);
-            int encHeight = even(height * scale);
+            int encWidth = even((int) (width * scale));
+            int encHeight = even((int) (height * scale));
 
             mOutputFile = RecorderPaths.newRecordingFile(mContext, format);
